requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-dotenv
httpx
pytest
pytest-asyncio
pandas
numpy>=1.24.0
scipy>=1.10.0
pandas-ta
aiosqlite
langchain==0.3.0
langchain-anthropic==0.3.15
langchain-openai==0.3.19
langchain-huggingface==0.2.0
langchain-community==0.3.0
langgraph==1.0.7
redis[hiredis]
tenacity
streamlit
plotly
streamlit-autorefresh
sentence-transformers
